add experimental settings, isolate run autonomous task
